add decorator for each endpoint

+++ api/views/todos_views.py
@@ -27,24 +27,36 @@ todo_args_schema = TodoArguments()
 
 
 @todos.route('/', methods=['GET'])
+@authenticate(token_auth)
 @arguments(todo_args_schema)
 @paginated_response('todos')
 def get_todos(args):
-    """Get todos"""
-    todos = ToDo.query
-    return todos
+    """Get todos
+
+    Query todo task for authenticated user. Only return back task created by the authenticated user. 
+    """
+    current_user = token_auth.current_user()
+
+    search = args.get('search', '')
+
+    task = db.session.query(ToDo).filter_by(user=current_user)
+
+    if search:
+        task = task.filter(ToDo.name.ilike(f'%{search}%'))
+
+    return task
 
 
-@todos.route('/', methods=['POST'])
-@response(todo_response_schema)
-@body(todo_entry_schema)
-@authenticate(token_auth)
+@ todos.route('/', methods=['POST'])
+@ response(todo_response_schema)
+@ body(todo_entry_schema)
+@ authenticate(token_auth)
 def create(args):
     """
     Create todo
 
-    This endpoint allow user to create todo. The target time represent the time the user 
-    wanted to use to accomplish the task. 
+    This endpoint allow user to create todo. The target time represent the time the user
+    wanted to use to accomplish the task.
     """
     user = token_auth.current_user()
     todo = ToDo(**args)
@@ -54,13 +66,13 @@ def create(args):
     return todo
 
 
-@todos.route('/details/<id>', methods=["GET"])
-@response(todo_response_schema)
-@authenticate(token_auth)
+@ todos.route('/details/<id>', methods=["GET"])
+@ authenticate(token_auth)
+@ response(todo_status_schema)
 def details(id):
     """Todo details
 
-    This route get the details of the todo associated with an authenticated user. 
+    This route get the details of the todo associated with an authenticated user.
     """
     current_user = token_auth.current_user()
     todo = ToDo.query.filter_by(id=id).first_or_404()
@@ -70,21 +82,22 @@ def details(id):
     return todo
 
 
-@todos.route('/update/<int:id>', methods=['PUT'])
-@response(todo_response_schema)
-@other_responses({200: "Task completed, cannot be update"})
-@body(todo_update_schema)
-@authenticate(token_auth)
+@ todos.route('/update/<int:id>', methods=['PUT'])
+@ authenticate(token_auth)
+@ response(todo_response_schema)
+@ other_responses({200: "Task completed, cannot be update"})
+@ body(todo_update_schema)
 def update(args, id):
     """Update todo
 
-    Authenticated user can update their todo by the id. 
+    Authenticated user can update their todo task by the id.
     """
     current_user = token_auth.current_user()
     todo = ToDo.query.filter_by(id=id).first_or_404()
 
     if todo.user_id != current_user.id:
         abort(401)
+        
     if todo and todo.is_completed == False:
         todo.update(args)
     elif todo.is_complete == True:
@@ -92,17 +105,17 @@ def update(args, id):
     return todo, 201
 
 
-@todos.route('/start/<id>', methods=['GET'])
-@authenticate(token_auth)
+@ todos.route('/start/<id>', methods=['GET'])
+@ authenticate(token_auth)
 def begin_todo(id):
-    """Start todo
+    """Start task
 
     Authenticated user can start their todo in thei todo list.
     """
     current_user = token_auth.current_user()
     todo = ToDo.query.filter_by(id=id).first_or_404()
 
-    if current_user.id != todo.id:
+    if current_user.id != todo.user_id:
         abort(401)
 
     if todo and todo.is_completed == True:
@@ -115,8 +128,19 @@ def begin_todo(id):
     return {}, 200
 
 
-@todos.route('/complete_todo/<id>', methods=['PUT'])
+@ todos.route('/complete_todo/<id>', methods=['PUT'])
+@ authenticate(token_auth)
 def complete(id):
+    """Complete task
+
+    Authicated user mark their todo task co
+    """
+    current_user = token_auth.current_user()
+    todo = ToDo.query.filter_by(id=id).first_or_404()
+
+    if todo.user_id != current_user.id:
+        abort(401)
+
     todo = ToDo.query.filter_by(id=id).first()
     if todo and todo.is_suspended == True:
         return jsonify({"status": "Task has suspended by you."})
@@ -132,8 +156,16 @@ def complete(id):
     return {}, 404
 
 
-@todos.route('/suspend_todo/<id>', methods=['PUT'])
+@ todos.route('/suspend_todo/<id>', methods=['PUT'])
+@ authenticate(token_auth)
 def suspend(id):
+    """Suspend task"""
+    current_user = token_auth.current_user()
+    todo = ToDo.query.filter_by(id=id).first_or_404()
+
+    if todo.user_id != current_user.id:
+        abort(401)
+
     todo = ToDo.query.filter_by(id=id).first()
     if todo and todo.is_completed == True:
         return jsonify({"status": "Task already completed, cannot be suspend."}), 200
@@ -145,8 +177,16 @@ def suspend(id):
     return {}, 404
 
 
-@todos.route('/delete_todo/<id>', methods=['DELETE'])
+@ todos.route('/delete_todo/<id>', methods=['DELETE'])
+@ authenticate(token_auth)
 def delete_todo(id):
+    """Delete task"""
+    current_user = token_auth.current_user()
+    todo = ToDo.query.filter_by(id=id).first_or_404()
+
+    if todo.user_id != current_user.id:
+        abort(401)
+
     todo = ToDo.query.filter_by(id=id).first()
     if todo:
         db.session.delete(todo)
